[bridges] Add unpolished bridge elim rule. add elim examples.
If a bridge P is given as well as a fresh bridge var i,
then P i typechecks as expected.
Since computation rules are missing, endpoints conditions
typically
result in  unsatisfied typechecking constraints. Hence it
is for now hard to tell if unfresh application would
typecheck as well (it should not).