[bridges] Make extent beta reduction less eager
Suppose we want to reduce extent ... r M where r:BI and
M is principal argument.
The principal argument M of extent is reduced (ie whnf M' ) before
the semi freshness analysis "r in M" occurs, as before.
But the reduction is now less eager because it does not use M' as
a result. It instead uses simplified M, this makes the output
more readable.
It also seems to fix a bug appearing when M is extent itself.

The above strategy is fine for now but I fear that it is
not complete! there probably are cases where we would like
reduction to be more eager instead.