[bridges] Fix and finish hasEmptyMeet. Test it with script.
`hasEmptyMeet zeta1 zeta2` determines if the intersection of
the mixed constraints zeta1 and zeta2 is vaccuous.
Precondition: zeta1, zeta2 path and bridge components are atomic,
i.e. are an intersection of path hyperplanes or i0 (in path case)
or are bridge hyperplanes or bottom BCstr (in bridge case).

`hasEmptyMeet` is useful for the coherence check arising when
typechecking something defined by patt. match. in  MPartial zeta A.
We have to check that the clause bodies "agree where they overlap".
When the overlap is empty, this holds trivially.