[bridges] Coherence check, wip
When typechecking a term mpart : MPartial zeta A
defined by pattern matching on the mixed constraint zeta,
in the coherence check,
we have to compare the bodies of rhs "where the lhs's overlap".
This is achived by pulling back types and bodies of rhs along
certain substitutions and checking judg equality in a shorter ctx.