[bridges] Add new primitives: mixed Kan operations. Pragma exp. wip.
primMComp (het comp), primMHComp (hom comp)
take as parameter a mixed partial  `u : (i:I) -> MPartial zeta A`,
instead of a partial element in the --cubical case.
The plan is to duplicate primComp, primHcomp, primTransp but
in this mixed setting, and ultimately to redefine the latter
--cubical primitives as instances of the former.

For the sake of independance betw. cubical and bridges, I'm
experimenting wrt the --cubical / --bridges flag exact
bahviour.
conjecture: enough to drop a few (inconsequential) --bridges
flag requirements in some bridges prims
in order to be able to define --cubical stuff
as those mixed versions.