[bridges] Flag downgrade --bridges -> --cubical for several prims.
Ultimately, the cubical Kan operations should be defined (reduce)
as instances of mixed Kan operations.
To build those terms `mixedKan (args)` we need the mixed Kan
operations to only requireCubical (instead of requireBridges, as
we naively expected). And to build the correct `args` we
need to downgrade several other primitives from --bridges
to --cubical as well. We list those primitives:

 - bdg interval fragm.:
     (BI : LockU), (bi0, bi1 : BI)
 -   (CstrUniv : UnivSort _)
 - BCstr fragm.:
     (BCstr : CstrUniv), (bno, byes : BCstr)
     ( _=bi0, _=bi1 : BI -> BCstr)
     (primBConj : BCstr -> BCstr -> BCstr)
     (BHolds : BCstr -> SSet) (BitHolds : BHolds byes)
 - MCstr fragment:
     (MCstr : CstrUniv) (mno, myes : MCstr)
     (primMkmc : I -> BCstr -> BCstr)
     (MHolds : MCstr -> SSet) (MitHolds : MHolds myes)
 - MPartial : {l : Level} -> MCstr -> Type l -> SSet l
 - Mixed Kan ops fragment (which are wip):
     reflectMCstr : {phi : I} -> .(MHolds phi m-or bno) ->IsOne phi
     primMHComp : ...
     primMComp : ...
     ...

Note that the extent and Gel fragments remain --bridges.
Note that I'm not sure downgrading all of them is really required.

However, since with this commit the --cubical user has access to
strictly more primitives, we have to make sure that she is not
able to prove new things.
This is done by having side conditions in mixed Kan ops,
which are the only primitives that could modify the expressive
power of cubical. side cond: is --bridges disabled then
current mixed constraint is in fact pure-path.