[bridges] Change (slow) reduce st HIT elim at pure mhocom reduces.
This is wip.

Functions defined by pattern matching on HIT members autom. get
computational rules for hcomp elements. Our long-term goal is
to generate similar reductions for mhocom elements.
But for now, we modify `slowReduceTerm` (todo: fast red) s.t.:
when we reduce a function f that has an hcomp clause (ie it
was defined by pattern matching on a HIT value), on a mhocom value
M that is pure-path (ie the mixed constraint is `embd (phi:I)`)
the we fold M to an hcomp.

This should allow us to typecheck files of the cubical library that
use HITs, with --bridges enabled.