[bridges] primRefoldMhocom, the come back
When _simplifying_ `primRefoldMhocom (mhocom ...zeta..)` you get
either a folded hcomp if zeta was path-pure, or the same mhocom
term if zeta was mixed.

Such a simplified term is now used when we match a function f that
has an hcomp clause on a mhocom value. This is useful
to typecheck --cubical HIT code with the --bridges flag enabled.
Indeed, if --bridges is enabled then hcomp->mhocom (although
this is not a simplification step).